Trait memory_x86_64::BootInformation

pub trait BootInformation: 'static {
    type MemoryRegion<'a>: MemoryRegion;
    type MemoryRegions<'a>: Iterator<Item = Self::MemoryRegion<'a>>;
    type ElfSection<'a>: ElfSection;
    type ElfSections<'a>: Iterator<Item = Self::ElfSection<'a>>;
    type Module<'a>: Module;
    type Modules<'a>: Iterator<Item = Self::Module<'a>>;

    fn start(&self) -> Option<VirtualAddress>;
    fn len(&self) -> usize;
    fn kernel_memory_range(
        &self
    ) -> Result<Range<PhysicalAddress>, &'static str>;
    fn bootloader_info_memory_range(
        &self
    ) -> Result<Range<PhysicalAddress>, &'static str>;
    fn modules_memory_range(
        &self
    ) -> Result<Range<PhysicalAddress>, &'static str>;
    fn memory_regions(&self) -> Result<Self::MemoryRegions<'_>, &'static str>;
    fn elf_sections(&self) -> Result<Self::ElfSections<'_>, &'static str>;
    fn modules(&self) -> Self::Modules<'_>;
    fn rsdp(&self) -> Option<PhysicalAddress>;
    fn stack_size(&self) -> Result<usize, &'static str>;
}

Required Associated Types

type MemoryRegion<'a>: MemoryRegion

type MemoryRegions<'a>: Iterator<Item = Self::MemoryRegion<'a>>

type ElfSection<'a>: ElfSection

type ElfSections<'a>: Iterator<Item = Self::ElfSection<'a>>

type Module<'a>: Module

type Modules<'a>: Iterator<Item = Self::Module<'a>>

Required Methods

fn start(&self) -> Option<VirtualAddress>

Returns the boot information’s starting virtual address.

fn len(&self) -> usize

Returns the boot information’s length.

fn kernel_memory_range(&self) -> Result<Range<PhysicalAddress>, &'static str>

Returns the range of physical addresses at which the kernel code is located.

fn bootloader_info_memory_range(
&self
) -> Result<Range<PhysicalAddress>, &'static str>

Returns the range of physical addresses at which the bootloader information is located.

fn modules_memory_range(&self) -> Result<Range<PhysicalAddress>, &'static str>

Returns the range of physical addresses at which the modules are located.

fn memory_regions(&self) -> Result<Self::MemoryRegions<'_>, &'static str>

Returns memory regions describing the physical memory.

fn elf_sections(&self) -> Result<Self::ElfSections<'_>, &'static str>

Returns the kernel’s ELF sections.

fn modules(&self) -> Self::Modules<'_>

Returns the modules found in the kernel image.

fn rsdp(&self) -> Option<PhysicalAddress>

Returns the RSDP if it was provided by the bootloader.

fn stack_size(&self) -> Result<usize, &'static str>

Returns the stack size in bytes.