Crate memory_x86_64

This crate implements the virtual memory subsystem interfaces for Theseus on x86_64. memory uses this crate to get the memory layout and do other arch-specific operations on x86_64.

This is the top-level arch-specific memory crate. All arch-specific definitions for memory system are exported from this crate.

Structs

BootInformation

A Multiboot 2 Boot Information struct.

EntryFlags

Page table entry flags on the x86_64 architecture.

Constants

PAGE_TABLE_ENTRY_FRAME_MASK

A mask for the bits of a page table entry that contain the physical frame address.

Functions

find_section_memory_bounds

Finds the addresses in memory of the main kernel sections, as specified by the given boot information.

get_boot_info_mem_area

Gets the physical memory area occupied by the bootloader information.

get_kernel_address

Finds and returns the relevant addresses for the kernel image loaded into memory by the bootloader.

get_modules_address

Gets the address bounds of physical memory occupied by all bootloader-loaded modules.

get_p4

Returns the current top-level page table address.

get_vga_mem_addr

Gets the physical memory occupied by vga.

tlb_flush_all

Flushes the whole TLB.

tlb_flush_virt_addr

Flushes the specific virtual address in TLB.